(PUP-10118) Add a method to check if a Puppet::Type is in the catalog
The `resource_refs` method returns the type and title of each resource in the
catalog. The title uniquely identifies the resource in the catalog, but the
title may not match the name of the resource as known by the system.

Sometimes we need to know if a `Puppet::Type` instance is contained in the RAL
catalog. For example when purging unmanaged resources, or establishing
autorelations.

The `resource_refs` method has two problems which make it unsuited for that use
case. First, it encourages a linear scan of the catalog, which is often times
repeated, leading to quadratic behavior. Second, it doesn't take into account
aliases, which are created when the resource's namevar(s) don't match its title.

This commit adds a `contains_resource?` method to check if a `Puppet::Type`
instance is contained in the catalog. It matches based on the
`[type, title]`, or the [type, uniqueness_key], where the uniqueness key
contains the values of each namevar. For the package type, the namevars are
`:command`, `:name`, `:provider`, so an example uniqueness key is
`[nil, 'httpd', :yum]`