(PUP-10118) Use contains_resource?
Call `Catalog#contains_resource?` to check if a resource is contained in the
catalog. This checks the resource type and title, as well as namevar(s), in the
case that the title doesn't match its name, or the resource has multiple
namevars. This also avoid scanning the entire catalog, once per purged resource
or unknown key in the ssh authorized keys file.