(maint) Lookup the PATH value case insensitively
Previously, the test was sensitive to the case of the PATH vs Path environment
variable.

When run in powershell (like appveyor), the variable has case `Path`. The call
to `withenv` correctly sets the value of the environment for the duration of the
block. However, `get_environment_strings` returns a hash whose keys may be mixed
case.

When running the test in cmd.exe, the variable has case `PATH`, so
`envhash['Path']` returned nil, and the test failed.

Update the test to preserve the case of the path environment variable when
verifying its value.