(...) Change all `on_write_only` hooks to `on_initialize_and_write`
By default hooks are only called with values specified in defaults, in the
`[main]` section or on the command line, but not in other sections.

Change settings with `on_write_only` hooks to `on_initialize_and_write` so they
are called when set in other sections, such as `[agent]`.

Since the `pluginsync` setting defaults to true, we only want to issue a warning
when not set in defaults.