(..) Delete on_write_only hook
Hooks of type `:on_write_only` are never called if the setting is contained in
`[agent]` or `[server]`, and that has bitten us many many times.

Delete the `:on_write_only` hook type. Make `:on_initialize_and_write` the
default. As a result of that change, the hook will be called with the default
value and after application defaults are initialized. Some hooks such as
`ssl_client_ca_auth` had to be updated so that validation is skipped if called
with the default value (so Puppet.settings.set_by_config? returns false).
Similarly for `masterport` and `serverport` (as the default values are 8140).