(maint) Globbing non-existent directories is expensive
The autoloader is responsible for loading applications, termini, types,
providers, report processors, etc. The autoloader scans the module path so that
these extensions can be provided in a module, however, that is rarely the case.
This commit checks to see if the directory exists before trying to globbing it.

It turns out globbing a non-existent directory is about 2x as expensive on
RHEL 7 and about 15x more expensive on Windows 2016.

On redhat 7:

    glob    348.532k (± 0.4%) i/s -      1.772M in   5.085259s
    stat    695.674k (± 0.7%) i/s -      3.496M in   5.025713s

On Windows, the difference is more extreme:

    glob      6.363k (± 7.1%) i/s -     31.742k in   5.017587s
    stat     98.034k (± 7.3%) i/s -    495.465k in   5.083427s