(maint) Lookup current environment
Previously the autoloader did not consistently use the current
environment to load code. As a result, it could mix code from different
environments. For example, when running as an agent with a "remote"
environment, it would start off only using the libdir, load a provider,
then try to evaluate a provider feature and lose track of the current
environment. It would then fall back to the configured environment,
and could load code from the modulepath.