(maint) Resolve Lint/MissingSuper
Explicitly call `super` to avoid bugs that can occur if functionality is added
to a superclass' initialize method, but isn't added to all of the subclasses
that don't call `super`.

There are a few places where it's not clear if calling super is the correct
thing, and in those cases the cop has been disabled.

In cases where the super and subclass initialize an instance variable, have the
subclass call to the superclass' initialize method, such as the
EvaluatingEppParser/EvaluatingParser, PSensitiveType/PAnyType, and HTTP client
resolvers.

Also call super for `inherited` and `method_added` lifecycle methods.