(PUP-9570) Check for mismatched environments
Previously, puppet agent could download plugins and facts from one
environment (A), but puppetserver could try to compile a catalog in a different
environment (B). If puppet code in B relied on a fact that only existed in B,
then compilation could fail, depending on how the fact was used. If fact
versions were different in A and B, then results were undefined.

If the agent requests the check be performed and there is an environment
mismatch, then short-circuit catalog compilation by returning an empty catalog
with the server specified environment. The agent will detect the mismatch and
retry its convergence loop, up to a maximum limit.

Note the agent won't cache the empty catalog, because caching only occurs if the
environment converges and it successfully converts it to a RAL catalog.