(PUP-10955) Always cache environments using symbols
The `add_entry` method cached environments based on the `name` argument, which
could be a string or symbol. Now we normalize environment names to symbols for
all public methods that resolve a name into an Environment, and we always use
the `Environment#name` when adding to or deleting environments from the cache.

The logging specs expected the environment to not reload, so set the timeout to
unlimited. Previously they passed due to string/symbol mismatch.