(PUP-11318) Use flat_map not +
Commit af560639d0 flattened the code produced by the code merger to avoid
a deeply nested tree, but it used += to concatenate the children.

Commit f2ae1b0404 switched to `+` which eliminated the Lint/UselessAssignment
rubocop warning, but still created a new Array containing a copy of the memo and
the parsed class' code. This meant every iteration copied the memo and the memo
increased in size each time.

This commit uses `flat_map`, which correctly appends a single element or an
array of elements to the `children` array and has the same memory profile as
calling concat (for arrays) or << (for an element).