(PUP-11328) Fallback to node request
If we can't load our last server specified environment, then fallback to making
a node request like we used to. This way if we're supposed to use a server
specified environment and we upgrade, then we don't revert to "production"
and redownload plugins.

The logic is similar to what was removed in commit 39df438ef except we're using
`fail_on_404` so `Puppet::Node.indirection.find` will raise instead of returning
nil.

We also don't clear facts/query_options.