(...) Allow all occurrences to be knocked out
Currently knockout_prefix only works when the key is present in 2 or
fewer levels in the hierarchy (i.e. a single merge). If the key is
added to a level in between those, then the value won't be knocked out.

This commit adds a knockout_all lookup option which will remove all
occurences of the key...