(PUP-11717) Protect against facter returning a Hash subclass
Facter.resolve v4 returns a Facter::FactCollection Hash subclass unlilke Facter
v3. This causes functions like `dig` to fail, since the function only
accepts a Collection type, not a Runtime type:

    dig($facts, "ruby", "version")

Protect against this by calling `to_h`. If the object is a Hash, it will return
`self`. If the object is derived from Hash, it will return a new Hash containing
the contents of the object.