(..) Serialize via YAML correctly
Loading a Puppet::Resource::Catalog from YAML (both safe and unsafe) resulted in
a partially deserialized class, because the parent class (SimpleGraph) included
the PsychSupport mdoule, but Catalog did not. This can be seen by trying to
roundtrip a Catalog instance variable that only exists in Catalog, but not
SimpleGraph.

As a result of this change, a catalog's name must be initialized in order to
roundtrip correctly.