Error when trying to ensure non-existent services are stopped
When managing the `ensure` state for a service, the provider queries the
service status to determine if it's running or not. However, if the
service doesn't exist, then it returns that the service is `stopped`.

When trying to ensure a non-existent service is stopped, puppet thinks
the service is "insync", so returns exit code 0.

This commit updates the systemd provider's enable method to return
either :running, :stopped or :absent. To differentiate between `stopped`
vs `absent`, the provider first checks if the service exists.

This adds `absent` as a valid value for the `ensure` property, however,
it can't be set on a provider, since we don't want to remove services:

    puppet resource service ufw ensure=absent
    Error: The systemd provider can not handle attribute ensure