(PUP-11937) Skip autorequire if exec param is lazily evaluated
If an exec's 'command' parameter is lazily evaluated and is deferred,
then we can't autorequire the file that the command will execute. This
because we won't know what the command is until the exec is evaluated.
But that means it's already too late to add a relationship from the file
resource to the exec resource.