Wrap dpkg commands
If an `environment` parameter is specified for a package resource, then
the dpkg provider will enable those environment variables in the current
process, so that they will be inherited by the child process.

Since the `environment` can only be specified in individual package
resources, we can only modify instance methods on the provider, not
class methods.