Check for nil before closing Uniquefile
On Windows, if squelch was false and Uniquefile.new raised, then we attempted to
close a nil stdout. This scenario can occur if TEMP or similar environment
variable is defined but refers to a non-existent directory.

So check for nil before closing. We don't have this issue on posix, because
it uses an IO pipe instead of Uniquefile.

Fixes #9385