Add task to generate configuration reference
    bundle exec rake ref:config

produces generated/configuration.md using `puppet doc -r configuration`

In order for the output to not be dependent on the hostname/domain facts, allow
them to be overridden as an environment variable. This can't be done using
FACTER_*, because structured facts like networking.hostname can't be overridden
by an env variable.