Support arrays of deferred parameter values
If an `exec` resource's `environment` parameter was specified as an
array of values, where at least one of the values was deferred, then
validation failed with:

    undefined method `=~' for #<Puppet::Pops::Evaluator::DeferredValue:0x00007ff6913f7908

This occurred because the transaction's `resolved_resolve` assumed the
parameter's value was single-valued. As a result, the deferred value
was never resolved before calling the type's `validate` method.

This updates the transaction to support multi-valued parameters. All
deferred values will be resolved before the type's `validate` method is
called.

Fixes #9438