Prefer is_a? instead of instance_of?
is_a? considers the object's class ancestors and is generally
recommended to future proof your code, as your code won't need to change
if a derived class in introduced in the future.

instance_of? only returns true if the object's class is an exact match
and is generally faster, because it doesn't need to walk the ancestor
class hierarchy.

We were using both is_a? and instance_of? to check for DeferredValue.
Since this code is not called during compilation, use is_a? consistently.