(PUP-8670) Display usage for external subcommands
When running `puppet help <app>` where `app` is not a legacy nor face
application, puppet would error with:

    Error: Could not load help for the face <app>

Now we check if there is a valid `puppet-<app>` command using Puppet::Util.which
If one is found, execute `puppet-<app> --help` and display the resulting output.

It's important to execute the dashed version of `puppet-<app>` and not `puppet
<app>` so that we call the external subcommand directly, in the same way
Puppet::Util::CommandLine does.