Don't convert environment parameter to an integer
When requesting a catalog from an environment whose name consisted of only
digits, the handler converted the value to an integer, and later attempts to
validate the env name against a regex failed.

    ... Puppet Server Error: undefined method `match' for 42:Integer
    /opt/puppetlabs/puppet/lib/ruby/vendor_ruby/puppet/node/environment.rb:127:in `valid_name?'

Since we know the value of the environment query parameter should always be a
string, special case it.